[Feat] Create Api register so system error (Pull code and fix)

+++ user-service/src/main/java/com/huypt/user_service/dtos/status/Message.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class Message {
     public String message;
-    public long status;
+    public int status;
 }
